refactor: Voucher 구현 수정
