+++ src/main/java/com/devcourse/springbootbasic/application/Platform.java
@@ -3,9 +3,9 @@ package com.devcourse.springbootbasic.application;
 import com.devcourse.springbootbasic.application.exception.InvalidDataException;
 import com.devcourse.springbootbasic.application.io.InputConsole;
 import com.devcourse.springbootbasic.application.io.OutputConsole;
-import com.devcourse.springbootbasic.application.template.CommandLineTemplate;
-import com.devcourse.springbootbasic.application.template.CreateMenuTemplate;
-import com.devcourse.springbootbasic.application.template.ListMenuTemplate;
+import com.devcourse.springbootbasic.application.util.template.CommandLineTemplate;
+import com.devcourse.springbootbasic.application.util.template.CreateMenuTemplate;
+import com.devcourse.springbootbasic.application.util.template.ListMenuTemplate;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
