+++ src/main/java/com/devcourse/springbootbasic/application/repository/customer/CustomerRepository.java
@@ -2,6 +2,7 @@ package com.devcourse.springbootbasic.application.repository.customer;
 
 import com.devcourse.springbootbasic.application.constant.YamlProperties;
 import com.devcourse.springbootbasic.application.exception.InvalidDataException;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Repository;
 
 import java.io.IOException;
