+++ src/main/java/com/devcourse/springbootbasic/application/util/factory/FixedVoucherFactory.java
@@ -1,17 +1,18 @@
-package com.devcourse.springbootbasic.application.factory;
+package com.devcourse.springbootbasic.application.util.factory;
 
-import com.devcourse.springbootbasic.application.domain.PercentDiscountVoucher;
+import com.devcourse.springbootbasic.application.domain.FixedAmountVoucher;
 import com.devcourse.springbootbasic.application.domain.Voucher;
 import com.devcourse.springbootbasic.application.dto.VoucherType;
 
 import java.util.UUID;
 
-public class PercentVoucherFactory implements VoucherFactory {
+public class FixedVoucherFactory implements VoucherFactory {
+
     @Override
     public Voucher create(double voucherDiscount) {
-        return new PercentDiscountVoucher(
+        return new FixedAmountVoucher(
                 UUID.randomUUID(),
-                VoucherType.PERCENT_DISCOUNT,
+                VoucherType.FIXED_AMOUNT,
                 voucherDiscount
         );
     }
